+++ rpio/__init__.py
@@ -7,7 +7,7 @@
 # * permission of Bert Van Acker
 # *************************************************************************
 
-VERSION = (0, 3, "4-b")
+VERSION = (0, 3, "5")
 __version__ = ".".join([str(s) for s in VERSION])
 
 __title__ = "robosapiensio"