+++ rpio/transformations/transformations.py
@@ -176,6 +176,32 @@ def message2code_py(system=None,path="output/generated/messages"):
         f.write(template.render(messages=system.messages))
 
 
+def robochart2aadlmessages(maplek=None,path="output/generated/messages"):
+    """Function to generate AADL messages from robochart models
 
+    :param [MAPLEK]: [MAPLE-K modeled within robochart], defaults to [None]
+    :type [MAPLEK]: [maplek (robochart)](, optional)
+
+    :param [path]: [path to the output folder], defaults to ["output/generated/messages"]
+    :type [path]: [string](, optional)
+
+    ...
+    :return: [Functions returns nothing]
+    :rtype: [None]
+    """
+
+    if not exists(path):
+        mkdir(path)
+
+    # Initialize the Templates engine.
+    this_folder = dirname(__file__)
+    jinja_env = jinja2.Environment(loader=jinja2.FileSystemLoader(this_folder), trim_blocks=True, lstrip_blocks=True)
+
+    # Load the template
+    template = jinja_env.get_template('templates/aadl_messages.template')
+
+    # Extract all processes from AADL system model
+    with open(join(path, "messages.aadl"), 'w') as f:
+        f.write(template.render(types=maplek.types))